test(torghut): split linked recovery postgres coverage
